requirements.txt
# =============================================================================
# requirements.txt — Face + QR Attendance System dependencies
# =============================================================================

# Core vision
opencv-python>=4.8.0
numpy>=1.24.0

# Liveness detection (blink / EAR) — replaces dlib, no compiler required
mediapipe>=0.10.0

# Face recognition — no compiler required, wraps multiple DNN backends
deepface>=0.0.93
onnxruntime>=1.17.0       # CPU inference backend used by DeepFace models

# QR code
qrcode[pil]>=7.4.2        # generation
pyzbar>=0.1.9             # decoding (needs libzbar DLL on Windows — see README)

# Database
mysql-connector-python>=8.3.0

# Optional — FastAPI endpoints
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
